share autocompletion components in common module
