+++ frontend/src/components/editor/expression-input.tsx
@@ -4,51 +4,33 @@
  * Expression Input - A single-line input that looks like shadcn/ui Input but has
  * template expression pill functionality backed by CodeMirror
  */
-import React, { useCallback, useMemo, useState } from "react"
+import React, { useCallback, useMemo } from "react"
+import { ActionRead } from "@/client"
 import { useWorkflow } from "@/providers/workflow"
 import { useWorkspace } from "@/providers/workspace"
-import {
-  autocompletion,
-  closeBrackets,
-  closeBracketsKeymap,
-  completionKeymap,
-} from "@codemirror/autocomplete"
-import {
-  history,
-  historyKeymap,
-  indentWithTab,
-  standardKeymap,
-} from "@codemirror/commands"
+import { closeBrackets } from "@codemirror/autocomplete"
+import { history } from "@codemirror/commands"
 import { bracketMatching, indentUnit } from "@codemirror/language"
 import { linter, type Diagnostic } from "@codemirror/lint"
 import { EditorState } from "@codemirror/state"
-import {
-  EditorView,
-  keymap,
-  placeholder,
-  type ViewUpdate,
-} from "@codemirror/view"
+import { EditorView, placeholder, type ViewUpdate } from "@codemirror/view"
 import CodeMirror from "@uiw/react-codemirror"
-import { AlertTriangle } from "lucide-react"
 import { useFormContext } from "react-hook-form"
 
 import { createTemplateRegex } from "@/lib/expressions"
 import { cn } from "@/lib/utils"
 import {
-  createActionCompletion,
   createAtKeyCompletion,
+  createAutocomplete,
   createBlurHandler,
+  createCoreKeymap,
   createExitEditModeKeyHandler,
   createExpressionNodeHover,
-  createFunctionCompletion,
-  createMentionCompletion,
   createPillClickHandler,
+  createPillDeleteKeymap,
   createTemplatePillPlugin,
-  createVarCompletion,
   editingRangeField,
   EDITOR_STYLE,
-  enhancedCursorLeft,
-  enhancedCursorRight,
   templatePillTheme,
 } from "@/components/editor/codemirror/common"
 
@@ -96,8 +78,6 @@ export interface ExpressionInputProps {
   placeholder?: string
   className?: string
   disabled?: boolean
-  showTypeWarnings?: boolean
-  onTypeConversion?: (originalValue: unknown, convertedValue: string) => void
   defaultHeight?: "input" | "text-area"
 }
 
@@ -107,27 +87,17 @@ export function ExpressionInput({
   placeholder: placeholderText = "Type @ to begin an expression...",
   className,
   disabled = false,
-  showTypeWarnings = true,
-  onTypeConversion,
   defaultHeight = "input",
 }: ExpressionInputProps) {
   const { workspaceId } = useWorkspace()
   const { workflow } = useWorkflow()
   const methods = useFormContext()
-  const [typeConversionWarning, setTypeConversionWarning] = useState<{
-    originalType: string
-    originalValue: unknown
-    convertedValue: string
-  } | null>(null)
-  const actions = workflow?.actions || []
+  const actions = workflow?.actions || ({} as Record<string, ActionRead>)
   const forEach = useMemo(() => methods.watch("for_each"), [methods])
 
   // Safe value conversion with error handling
   const safeValue = useMemo(() => {
     try {
-      // Reset warning state
-      setTypeConversionWarning(null)
-
       // Handle null/undefined
       if (value == null) {
         return ""
@@ -157,43 +127,26 @@ export function ExpressionInput({
         convertedValue = String(value)
       }
 
-      // Set type conversion warning
-      if (showTypeWarnings && originalType !== "string") {
-        setTypeConversionWarning({
-          originalType,
-          originalValue: value,
-          convertedValue,
-        })
-
-        // Call optional callback
-        onTypeConversion?.(value, convertedValue)
-
-        // Log warning for development
-        console.warn(
-          `ExpressionInput: Non-string value of type "${originalType}" was converted to string:`,
-          { original: value, converted: convertedValue }
-        )
-      }
-
       return convertedValue
     } catch (error) {
       console.error(
         "ExpressionInput: Failed to convert value to string:",
         error
       )
-      setTypeConversionWarning({
-        originalType: typeof value,
-        originalValue: value,
-        convertedValue: "[Conversion Error]",
-      })
       return "[Conversion Error]"
     }
-  }, [value, showTypeWarnings, onTypeConversion])
+  }, [value])
 
   const extensions = useMemo(() => {
     const templatePillPluginInstance = createTemplatePillPlugin(workspaceId)
 
     return [
+      // Keymaps
+      createPillDeleteKeymap(), // This must be first to ensure that the delete key is handled before the core keymap
+      createCoreKeymap(),
+      createAtKeyCompletion(),
+      createExitEditModeKeyHandler(),
+
       // Core setup
       history(),
       EditorState.allowMultipleSelections.of(true),
@@ -202,35 +155,13 @@ export function ExpressionInput({
       // Linting
       linter(expressionLinter),
 
-      // Keymaps
-      createExitEditModeKeyHandler(),
-      keymap.of([
-        {
-          key: "ArrowLeft",
-          run: enhancedCursorLeft,
-        },
-        {
-          key: "ArrowRight",
-          run: enhancedCursorRight,
-        },
-        ...closeBracketsKeymap,
-        ...standardKeymap,
-        ...historyKeymap,
-        ...completionKeymap,
-        indentWithTab,
-      ]),
-      createAtKeyCompletion(),
-
       // Features
       bracketMatching(),
       closeBrackets(),
-      autocompletion({
-        override: [
-          createMentionCompletion(),
-          createFunctionCompletion(workspaceId),
-          createActionCompletion(Object.values(actions).map((a) => a)),
-          createVarCompletion(forEach),
-        ],
+      createAutocomplete({
+        workspaceId,
+        actions,
+        forEach,
       }),
 
       // Placeholder
@@ -337,38 +268,6 @@ export function ExpressionInput({
           className={EDITOR_STYLE}
         />
       </div>
-
-      {/* Type conversion warning */}
-      {typeConversionWarning && (
-        <div className="mt-2 flex items-start gap-2 rounded-md border border-amber-200 bg-amber-50 p-3 text-sm">
-          <AlertTriangle className="mt-0.5 size-4 shrink-0 text-amber-600" />
-          <div className="flex-1">
-            <div className="font-medium text-amber-800">
-              Data type converted
-            </div>
-            <div className="mt-1 text-amber-700">
-              Received {typeConversionWarning.originalType} value, automatically
-              converted to string for editing.
-            </div>
-            {(typeConversionWarning.originalType === "object" ||
-              typeConversionWarning.originalType === "boolean" ||
-              typeConversionWarning.originalType === "number") && (
-              <div className="mt-2 text-xs text-amber-600">
-                <div className="font-medium">Original value:</div>
-                <div className="mt-1 max-w-md overflow-auto rounded bg-amber-100 p-1 font-mono">
-                  {typeConversionWarning.originalType === "object"
-                    ? JSON.stringify(
-                        typeConversionWarning.originalValue,
-                        null,
-                        2
-                      )
-                    : String(typeConversionWarning.originalValue)}
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
     </div>
   )
 }
